feat(core)!: Remove `enableLogs` and `beforeSendLog` experimental options (#17063)
When we first shipped support for
https://docs.sentry.io/product/explore/logs/ in the JavaScript SDK we
gated the logging related options (`enableLogs`, `beforeSendLog`) behind
`_experiments` namespace.

See https://github.com/getsentry/sentry-javascript/pull/15717 and
https://github.com/getsentry/sentry-javascript/pull/15814 for the PRs
that made this change.

For v10, we want to not make this experimental anymore, so we changed
the logging options to become top-level options to `Sentry.init`.