feat(node): Add `shouldHandleError` option to `fastifyIntegration` (#16845)
Resolves: https://github.com/getsentry/sentry-javascript/issues/16819

This requires the Fastify integration to be manually added while
initialising Sentry.

```typescript
Sentry.init({
  integrations: [
    Sentry.fastifyIntegration({
      shouldHandleError(_error, _request, reply) {
        return reply.statusCode >= 500;
      },
    });
  },
});
```

If `shouldHandleError` is provided in `setupFastifyErrorHandler`, it
overrides `shouldHandleError` provided from `fastifyIntegration`. So
this should not break any existing (pre-DC) usage.

In summary: 
- When `setupFastifyErrorHandler` is not used, `shouldHandleError` at
`Sentry.init` will be used.
- When `setupFastifyErrorHandler` is used but without
`shouldHandleError`, `shouldHandleError` at `Sentry.init` will be used.
- When in both `setupFastifyErrorHandler` and `Sentry.init` are used
with `shouldHandleError`, the one in `setupFastifyErrorHandler` is used.

Works for all Fastify versions supported (v3, v4, v5)

Note: E2E tests for this now requires to test overrides, so they need to
be run twice with both overriden and non-overriden applications. Not the
prettiest solution, but I could not figure out an alternative.