fix(node): Ensure tool errors for `vercelAiIntegration` have correct trace connected (#17132)
This fixes the problem that for tool errors triggered from vercel-ai
integration, traces were not connected.
This seemed to happen because errors bubble up to the global unhandled
rejection handler, where the span is no longer active and thus the trace
cannot be connected.

This PR fixes this by attaching the active span to the error and using
this in the unhandled rejection handler, if it exists.

Closes https://github.com/getsentry/sentry-javascript/issues/17108