feat(browser)!: Remove FID web vital collection (#17076)
This PR removes the FID web vital collection logic from the browser SDK.
This is a behaviour breaking change and should not be backported to
`v9`.

For reasons see #16560 but the TLDR is that FID has long been deprecated
and replaced by INP.

Added a note to the changelog with possible implications on the product.
Also thought of a minor edge case where the removal could have an effect
on SDK code: If a filtering decision in `beforeSend` depends on the
presence of the FID measurement. A very niche case IMHO but probably
good to mention anyway.

closes #16560