test(node): Replace express v5 integration tests with E2E test (#17147)
We used to duplicate all the express integration tests and run yarn
separately for them to test express v5. This is a bit unfortunate and
also hard to keep in sync (already we added/changed tests in express v4
but not v5). This PR replaces the express v5 stuff with a dedicated
express-v5 e2e test - we actually did not have any of that so far :O

I thought about using v5 by default and having an express-v4 e2e test
app instead, but express v4 is still vastly more in use than v5, so
that's probably easier.