fix(v9/astro): Revert Astro v5 storing route data to `globalThis` (#17185)
To share the route data from build-time with the server runtime, the
data was written on `globalThis`. However, this resulted in actually
modifying the server config of the working directory and not the build
directory (`dist` folder).

This PR reverts this change for now, resulting in not having all server
requests parametrized. We still do have some parametrization, as route
parametrization was recently improved for Astro v4 as well (without the
`'astro:routes:resolved'` hook)

Fixes https://github.com/getsentry/sentry-javascript/issues/17179