fix(astro): Construct parametrized route during runtime (#17190)
For v9, the changes were already reverted in this PR
(https://github.com/getsentry/sentry-javascript/issues/17179) to create
a quick fix we can release soon. However, this removed some
parametrization.

This PR here not only fixed the problem with continuously writing to
`globalThis` to share build-time data with the runtime (we don't do this
anymore). The route parametrization now happens only during runtime, as
we have access to the route segments at runtime with Astro v5.

This adds a **little** performance overhead when compared with the
previous approach (the route segments are now constructed during
runtime) - but this is not an expensive operation. The `.find` method
was used in the previous approach as well.


Fixes https://github.com/getsentry/sentry-javascript/issues/17179