ref(feedback): Remove almost duplicate `sendFeedback` rejection message (#17297)
Noticed this by chance, this is a redundant message (we also use this as
general fallback anyhow).

The two messages have been:

```
Unable to send Feedback. This is because of network issues, or because you are using an ad-blocker.
Unable to send Feedback. This could be because of network issues, or because you are using an ad-blocker
```

Which are virtually identical. We can safe some bytes there IMHO.