fix(replay): Call `sendBufferedReplayOrFlush` when opening/sending feedback (#17236)
I noticed that our replays w/ feedbacks have gaps while the user is
entering in their feedback. This is because we call `replay.flush` (not
to be confused w/ the public replay integration integration which also
has a flush) which only flushes the buffer and continues to buffer. What
we want is to call `sendBufferedReplayOrFlush` which will convert the
buffered replay into a session replay (after flushing) in order to
continue recording.