fix(browser): Ensure IP address is only inferred by Relay if `sendDefaultPii` is `true` (#17364)
This PR fixes a long-standing problem in the SDK where it would set
incorrect information about when Relay should (not) infer IP addresses
from sent envelope requests.

Previously, this was thought to be controlled by setting
`event.user.ip_address: '{{auto}}'`. However, after an incident in
Relay, it was determined that this is in fact not a reliably way to
control IP inference. Instead, SDKs should set
`event.sdk.settings.infer_ip: 'auto' | 'never'` (see closes
https://github.com/getsentry/sentry-javascript/issues/16252).
Unfortunately, this wasn't implemented immediately but is now taken care of
in this PR.