feat(node): Capture `SystemError` context and remove paths from message (#17331)
resolves https://github.com/getsentry/sentry-javascript/issues/17332

Node.js has a
[`SystemError`](https://nodejs.org/api/errors.html#class-systemerror)
error class.

This error type contains numerous properties which can be useful for
debugging but it also includes many of them in the error message which
stops fingerprinting from grouping these errors.

This PR adds a default enabled integration that copies the error
properties to a new context and strips paths from the error message.

If the top-level `sendDefaultPii` is not enabled, the `path` and `dest`
properties will not be included.