test(node): Ensure runner does not create unused tmp files (#17392)
Noticed locally that I sometimes got unexpected un-removed `tmp_xxx`
dirs in node-integration tests. After some debugging, I figured out the
reason: We clean the tmp dir up in `afterAll` of the test. However, if
all tests inside are skipped, `afterAll` is not executed. So if you run
tests in a filtered way - e.g. `yarn test -t "xxx"` - it would create
the tmp dirs, but not delete them.

This PR fixes this by moving the creation of the tmp dir into
`beforeAll`, which equally is not executed when all tests are skipped.