fix(node): Fix preloading of instrumentation (#17403)
Extracted out of
https://github.com/getsentry/sentry-javascript/pull/17371

I noticed that we were not fully consistent in instrumentation IDs for
integrations that have multiple instrumentation. The intent is that
users can provide the _integration name_ (e.g. `Http`) and it will
preload all http instrumentation. To achieve this, I adjusted the
preload filter code to look for exact matches as well as
`startsWith(`${name}.id`)`. I also adjusted the test to be more
declarative and mock/reset stuff properly (this lead to issues in the
linked PR, and should generally be a bit cleaner).

I also updated all instrumentation IDs to follow this pattern. We should
be mindful of following this with new instrumentation we add.