feat(cloudflare): Add `instrumentPrototypeMethods` option to instrument RPC methods for DurableObjects (#17424)
Fixes `The RPC receiver does not implement the method "METHOD_NAME
error` errors by wrapping methods and putting them back on the
prototype.

Added an option `instrumentPrototypeMethods` to opt into wrapping RPC
methods. These are potentially expensive to wrap because each invocation
wraps so we provide an option to either enable this as a whole or pass
method names that should be wrapped.

Also removes the flag to not record spans because no spans were
collected ever. I think this is related to connected traces not working
correctly for the SDK? cc @AbhiPrasad.

Closes: #17127