feat(browser): Add replay.feedback CDN bundle (#17496)
I think we originally decided to only have a error + feedback bundle and
a error + tracing + replay + feedback bundle for simplicity. But I
suppose it does not really make things easier that there is exactly one
(error + replay + feedback) missing, so I'd just add this here now. Esp.
if we want to add this in the loader this will be much easier to handle.