fix(browser): Ensure source is set correctly when updating span name in-place in `beforeStartSpan` (#17501)
When using `beforeStartSpan` in `browserTracingIntgration`, we document
that users should return a new options object rather than mutating the
passed in object in-place. However, nothing keeps users from setting the
name in-place like so:'

```js
browserTracingIntegration({
  beforeStartSpan: opts => {
    opts.name = 'changed';
    return opts;
  },
}),
```

In such a case, we previously didn't catch the mutation and wouldn't
adjust the span source attribute to `'custom'` accordingly. This PR
ensures we take care of this case, too. (found while writing a test for
#17500)