feat(node): Update `httpIntegration` handling of incoming requests (#17371)
This PR updates the handling of the Node SDK of incoming requests.
Instead of relying on @opentelemetry/instrumentation-http for this, we
now handle this internally, ensuring that we can optimize performance as
much as possible and avoid interop problems. This will also allow us to
extract this out of an OTEL instrumentation, eventually paving the way
for a non-OTEL SDK with basic tracing capabilities.

This change should not affect users, unless they are relying on very
in-depth implementation details. Importantly, this also drops the
`_experimentalConfig` option of the integration - this will no longer do
anything.
Finally, you can still pass
`instrumentation.{requestHook,responseHook,applyCustomAttributesOnSpan}`
options, but they are deprecated and will be removed in v11. Instead,
you can use the new `incomingRequestSpanHook` configuration option if
you want to adjust the incoming request span.

I have tried to ensure the spans look as similar as possible as before,
this should be reflected in the tests.

---------

Co-authored-by: Lukas Stracke <lukas.stracke@sentry.io>