fix(astro): Ensure traces are correctly propagated for static routes (#17536)
This refactors the astro middleware a bit to be more correct & future
proof.

Right now, http.server spans are not emitted by the node http
integration because import-in-the-middle does not work with Astro. So we
emit our own. This PR makes astro ready to also handle the base
http.server spans and enhance them with routing information.

It also handles static routes better: these do not emit spans anymore
now, and do not continue traces, but instead only propagate the
parametrized route to the client, no trace data.

One fundamental problem remains that is not fixed in this PR:
`Sentry.init()` is only injected via `page-ssr` into SSR pages. This
means that only once any SSR page is hit, the server-side part of Sentry
is initialized. If you hit a prerendered (static) page first, sentry
will not be initialized and thus neither errors are captured nor spans
created. For regular prerendered pages this should be fine because we do
not need to run anything at runtime there. However, when you hit a
prerendered page that has a server-island, the server island (which is
dynamic) will not be instrumented either because Sentry is not
initialized yet.

I don't think we can fix this with the current set of primitives we
have, so this is a future problem to look into...

---------

Co-authored-by: Lukas Stracke <lukas.stracke@sentry.io>