fix(browser): Ensure propagated `parentSpanId` stays consistent during trace in TwP mode (#17526)
This PR adjusts how we set treat the `parentSpanId` when propagating it
in an SDK configured for [tracing without
performance](https://develop.sentry.dev/sdk/telemetry/traces/tracing-without-performance).
Previously, we didn't set the `propagationSpanId` on the
`propagationContext` (which is supposed to be serialized to the parent
span id in the `sentry-trace` (and `traceparent`) header, if no active
span is present.
This had the consequence that every time `getTraceData()` was called,
we'd return a new, random (non-existing) parent span id. Meaning, that
if we generate both, `sentry-trace` and `traceparent` headers via
`getTraceData()` they'd end up with different parent span ids in the
respective hedaers.

With this change, we now set a `propagationSpanId` on the PC, which has
two effects:
1. consistent parent span id in `sentry-trace` and `traceparent` headers
for the same outgoing request
2. consistent parent span ids in all outgoing requests until the next
navigation when the PC is recycled

So this prolongs the time span in which we propagate the same parent
span id. Since this span doesn't exist, I _think_ this is fine. Though
happy to drop this if reviewers have concerns. The main benefit with
this change is 1 in combination with us generating `traceparent` from
raw data rather than parsing the already generated `sentry-trace` header
and converting its parts into `traceparent`.

(TwP was a mistake)