fix(nextjs): Add edge polyfills for nextjs-13 in dev mode (#17488)
This issue is only observed for Nextjs 13 and specifically for apps that
don't use the app router. Similarly to what we do in vercel-edge, we
polyfill `performance`. This is only done for dev mode as it's not an
issue otherwise.

Note: I tried to reproduce this via our existing e2e test apps but
they're all using the app router so it's not manifesting there.

Closes: #17343