feat(browser): Add timing and status atttributes to resource spans (#17562)
Adds a few attributes to `resource.*` spans for request timing
and status information. Most importantly:

- `http.request.time_to_first_byte` [from
`responseStart`](https://developer.mozilla.org/en-US/docs/Web/API/PerformanceResourceTiming/responseStart)
which can be [interpreted as
TTFB](https://web.dev/articles/ttfb#measure-resource-requests) for
resource requests. This was requested via via
https://github.com/getsentry/sentry/issues/63739.
- `http.response.status_code` the status code of the resource request.
Requested in #16805 and #10995

To get these attributes, I adjusted the already existing
`resourceTimingToSpanAttributes` a bit:
- Moved it from browser to browser-utils because resource spans logic is
located in browser-utils (this is safe, it never was exported from
browser)
- Changed the signature to return an object instead of an array of
attributes. This is more ergonomical to use in both callsites and should
reduce bundle size slightly
- Added `http.request.redirect_end`, `http.request.worker_start`
attributes which are stable timing values available on
`PerformanceResourceTiming` but were previously missing
- Also added to conventions:
https://github.com/getsentry/sentry-conventions/pull/130
- Added `http.request.time_to_first_byte`. Decided to add this attribute
because `http.request.response_start` cannot be directly used as TTFB as
its value is an absolute time stamp of `responseStart`. Instead the TTFB
attribute is the relative response start attribute converted to seconds
(happy to change to ms if reviewers prefer).
- Also added to conventions:
https://github.com/getsentry/sentry-conventions/pull/131

Other consequences:
- `http.client` spans now also have the three additional attributes

Remarks:
- Not super happy about us always defaulting to `0` in case a value is
not present. But decided to leave this as-is to avoid any behaviour
change for http.client spans.