ref(core): Streamline event processor handling (#17634)
Avoid re-creating many sync promises where not needed.

We used to wrap every result of an event processor in a sync promise, no
matter if needed or not. This is not really necessary, likely having a
(very minor) perf impact but also making the stacktraces etc. more
complicated then neccessary.

With this PR, we only wrap the overall result in a sync promise once, if
needed, and otherwise just run the event processors directly.

Somewhat inspired by looking into this for
https://github.com/getsentry/sentry-javascript/issues/17592, but likely
does not actually fix this.