fix(core): Fix error handling when sending envelopes (#17662)
I noticed that we actually handled errors in `sendEnvelope` incorrectly
- we would resolve this function with the rejection reason, if sending
fails. this does not match the type of `TransportMakeRequestResponse`,
you could actually get something like this out (and the tests actually
incorrectly tested this):

```js
// transport.send() rejects with "fetch does not exist"
const res = await client.sendEnvelope(envelope);
// res --> "fetch does not exist" (string)
```

This PR fixes this to instead resolve with an empty object (which
matches the expected return type).

Extracted this out of
https://github.com/getsentry/sentry-javascript/pull/17641 because it is
actually a bug/fix.