feat(core): Allow to pass `onSuccess` to `handleCallbackErrors`
I've seen a few places where we are wrapping things that could be sync or async, and we want to do something with the return value. By adding an onSuccess handler to `handelCallbackErrors` we can handle this more generically in the future.
revert and fix
vendor type in

better awaited??