feat(node): Do not drop 300 and 304 status codes by default (#17686)
Looking at
https://developer.mozilla.org/en-US/docs/Web/HTTP/Reference/Status#redirection_messages,
it seems that we should not filter out 300, 302 and 304 status codes by
default:

> [300 Multiple
Choices](https://developer.mozilla.org/en-US/docs/Web/HTTP/Reference/Status/300)
In [agent-driven content
negotiation](https://developer.mozilla.org/en-US/docs/Web/HTTP/Guides/Content_negotiation#agent-driven_negotiation),
the request has more than one possible response and the user agent or
user should choose one of them. There is no standardized way for clients
to automatically choose one of the responses, so this is rarely used.
> [304 Not
Modified](https://developer.mozilla.org/en-US/docs/Web/HTTP/Reference/Status/304)
This is used for caching purposes. It tells the client that the response
has not been modified, so the client can continue to use the same
[cached](https://developer.mozilla.org/en-US/docs/Web/HTTP/Guides/Caching)
version of the response.

In contrast, the others seem safe to continue to filter:

> [301 Moved
Permanently](https://developer.mozilla.org/en-US/docs/Web/HTTP/Reference/Status/301)
The URL of the requested resource has been changed permanently. The new
URL is given in the response.

> [302
Found](https://developer.mozilla.org/en-US/docs/Web/HTTP/Reference/Status/302)
This response code means that the URI of requested resource has been
changed temporarily. Further changes in the URI might be made in the
future, so the same URI should be used by the client in future requests.

> [303 See
Other](https://developer.mozilla.org/en-US/docs/Web/HTTP/Reference/Status/303)
The server sent this response to direct the client to get the requested
resource at another URI with a
[GET](https://developer.mozilla.org/en-US/docs/Web/HTTP/Reference/Methods/GET)
request.

> [305 Use Proxy
Deprecated](https://developer.mozilla.org/en-US/docs/Web/HTTP/Reference/Status#305_use_proxy)
Defined in a previous version of the HTTP specification to indicate that
a requested response must be accessed by a proxy. It has been deprecated
due to security concerns regarding in-band configuration of a proxy.

> [306
unused](https://developer.mozilla.org/en-US/docs/Web/HTTP/Reference/Status#306_unused)
This response code is no longer used; but is reserved. It was used in a
previous version of the HTTP/1.1 specification.

> [307 Temporary
Redirect](https://developer.mozilla.org/en-US/docs/Web/HTTP/Reference/Status/307)
The server sends this response to direct the client to get the requested
resource at another URI with the same method that was used in the prior
request. This has the same semantics as the 302 Found response code,
with the exception that the user agent must not change the HTTP method
used: if a
[POST](https://developer.mozilla.org/en-US/docs/Web/HTTP/Reference/Methods/POST)
was used in the first request, a POST must be used in the redirected
request.

> [308 Permanent
Redirect](https://developer.mozilla.org/en-US/docs/Web/HTTP/Reference/Status/308)
This means that the resource is now permanently located at another URI,
specified by the
[Location](https://developer.mozilla.org/en-US/docs/Web/HTTP/Reference/Headers/Location)
response header. This has the same semantics as the 301 Moved
Permanently HTTP response code, with the exception that the user agent
must not change the HTTP method used: if a
[POST](https://developer.mozilla.org/en-US/docs/Web/HTTP/Reference/Methods/POST)
was used in the first request, a POST must be used in the second
request.

eventually it makes sense to unify this between node/node-core, but this
should happen when we merge the httpIntegration split PR.