fix(browser): Ensure idle span duration is adjusted when child spans are ignored (#17700)
This PR fixes a problem reported in
https://github.com/getsentry/sentry-javascript/issues/17451 where
ignoring spans in idle root spans (pageload and navigations most
prominently) caused the root span duration to be perceived much longer
than reasonable. This is only a problem for idle spans in browser, so
this PR applies a pragmatic fix to adjust the end time stamp:

We already adjust the time stamp when we end the idle span. So we might
as well at this point take any to-be-removed-because-of-`ignoreSpans`
spans out of this calculation. This should work well enough without
throwing a bunch of idle-span specific logic into the client or
completely changing the point in the event lifecycle where `ignoreSpans`
is applied.

closes https://github.com/getsentry/sentry-javascript/issues/17451