ref(core): Avoid looking up anthropic-ai integration options (#17694)
This avoids looking up the integration for anthropic-ai instrumentation,
instead relying on this being passed in (which it already is). When
manually instrumenting the client you need to pass in the options
directly.

1. Node: Options are passed from the integration to
`instrumentAnthropicAiClient` anyhow, so nothing changes
2. cloudflare/vercel-edge: There is no integration, users need to
manually call `instrumentAnthropicAiClient()` and pass in the options
anyhow (no integration to look anything up from exists there)

This required updating the tests to actually use auto instrumentation
properly instead of manual client wrapping, which is overall a good
change anyhow IMHO.