ref(core): Wrap scopes in `WeakRef` when storing scopes on spans
We often store scopes on spans via `setCapturedScopesOnSpan` and retrieve them
via `getCapturedScopesOnSpan`. This change wraps the scopes in `WeakRef` to
attempt fixing a potential memory leak when spans hold on to scopes
indefinitely.

The downside is scopes might end up with undefined scopes on them if the scope
was garbage collected, we'll have to see if that's an issue or not.