ref(core): Wrap isolationscope in `WeakRef` when storing it on spans (#17712)
We often store isolationscopes on spans via `setCapturedScopesOnSpan` and
retrieve them via `getCapturedScopesOnSpan`. This change wraps the
scopes in `WeakRef` to attempt fixing a potential memory leak when spans
hold on to scopes indefinitely.

The downside is spans might end up with undefined scopes on them if the
scope was garbage collected, we'll have to see if that's an issue or
not.