feat(browser): Add `setActiveSpanInBrowser` to set  an active span in the browser (#17714)
This PR adds a long-requested feature to the browser SDKs (only!):
Making an inactive span active. We do this to enable use cases where
having a span only being active in the callback is not practical (see
https://github.com/getsentry/sentry-javascript/issues/13495 for
examples).

There are a couple of caveats to this feature:
- This on purpose is only exported from the browser SDKs. We cannot
support this in Node, due to OTel not allowing for a similar hack.
Frankly, it also makes no sense in Node-based SDKs.
- Calling `setActiveSpanInBrowser ` on a nested child span, will make
that child span the active span as long as it is active. However, due to
`parentSpanIsAlwaysRootSpan` defaulting to `true` any child span of the
active child span, will still be parented to the root span. By setting
`parentSpanIsAlwaysRootSpan: false`, the span hierarchy is respected and
child spans are correctly parented to the active span. Note that this
cannot be guaranteed to work perfectly, due to missing async context in
the browser. See tests for the `parentSpanIsAlwaysRootSpan` behaviour.
- A span once set active, cannot be set as inactive again. It will
remain active until it is ended or until another span is set active. In
the latter case, once that span ends, the initial span will be set as
active again until it ends. This is reflected in the types where we by
design only allow `Span` to be passed to `setActiveSpanInBrowser `.

Technically, `setActiveSpanInBrowser` uses `_setSpanForScope` which I
decided to re-export from core as `_INTERNAL_setSpanForScope`, similarly
to how we do it with logs APIs.

### Usage

This example shows one of the most frequent use cases where having an
active, callback-unbound span is useful:

```js
function instrumentMyRouter() {
  let routeSpan;
  
  on('routeStart', (from, to) => {
    routeSpan = Sentry.startInactiveSpan({name: `/${from} -> /${to}`});
    Sentry. setActiveSpanInBrowser(rootSpan);
  });


  // any span started in the meantime (e.g. fetch requests) will be
  // automatically parented to `routeSpan`
  
  on('routeEnd', () => {
    // automatically removes the span from the scope
    routeSpan.end();
  })
}
```

closes https://github.com/getsentry/sentry-javascript/issues/13495