feat(cloudflare): Add `honoIntegration` with error-filtering function (#17743)
This adds a `honoIntegration` that exposes a `shouldHandleError`
function that lets users define a custom handling of capturing errors.
By default, we capture exceptions with `error.status >= 500 ||
error.status <= 299`.

It's possible to modify this behavior like this:

```js
integrations: [
  honoIntegration({
    shouldHandleError: (err) => true; // always capture exceptions in onError
  })
]
```

**Up for discussion**: The function name is the same like in
express/fastify to keep it consistent. But I was also thinking about
other names like `shouldHandleInOnError` or `shouldCaptureOnError` 🤔
Because it's specifically about the `onError` function. Let me know what
you think.

closes https://github.com/getsentry/sentry-javascript/issues/17717