ref(aws-serverless): Add resolution for `import-in-the-middle` when building the Lambda layer (#17780)
this is actually a fix but not a user-facing one, therefore, `ref`.

Starting with `1.14.3`, `import-in-the-middle` uses both, `import` and
`require` statements in its `hook.mjs` file. This caused `@vercel/nft`
in its default config to _not_ trace the required `hook.js` file (+ all
of its dependencies). This PR enables `mixedModules: true` to trace
both, requires and imports. In contrast to #17777, we therefore don't
have to override the IITM version and we also don't have to revert this
PR.