ref(core): Improve promise buffer (#17788)
Extracted this out of
https://github.com/getsentry/sentry-javascript/pull/17782, this improved
our promise buffer class a bit:

1. Remove the code path without a `limit`, as we never use this (there
is a default limit used). There is also really no reason to use this
without a limit, the limit is the whole purpose of this class.
2. Use a `Set` instead of an array for the internal buffer handling,
this should slightly streamline stuff.
3. For `drain`, we can simplify the implementation without a timeout
drastically. We can use `Promise.race()` to handle this more gracefully,
which should be supported everywhere.
4. Some slight refactorings, actually improving timing semantics
slightly.