fix(nextjs): Don't use chalk in turbopack config file (#17806)
We don't know yet why but this config code seems to be evaluated once at
runtime for some users when using turbopack. The chalk import breaks
with `Failed to load external module @sentry/nextjs: Error
[ERR_REQUIRE_ESM]`.
Just getting rid of this import for now to not break apps.

We're still investigating together with Vercel why this this code gets
pulled in.

closes https://github.com/getsentry/sentry-javascript/issues/17691