fix(browser): Use current start timestamp for CLS span when CLS is 0 (#17800)
When there's no layout shift, the standalone CLS span would previously
have the same start time as the pageload span
(`performance.timeOrigin`). This caused the cls span to be incorrectly
interpreted as the trace root in the Sentry UI. We can fix this by
setting the span start timestamp as the startTime instead if there's no
layout shift. In all other cases, we let the span start at the time the
last CLS update occured.